Reorganized AWESim_SOSS into pipeline
