+++ AWESim_SOSS/wasp107_data/WASP107_LDC.txt
@@ -1,0 +1,91 @@
+Profile: quadratic
+|wavelength|    c1|    e1|    c2|    e2|
+|    double|double|double|double|double|
+|          |      |      |      |      |
+|      null|  null|  null|  null|  null|
+     0.8095  0.501  0.006  0.131  0.009 
+     0.8295  0.483  0.007  0.140  0.009 
+     0.8495  0.448  0.007  0.143  0.010 
+     0.8695  0.434  0.007  0.145  0.010 
+     0.8895  0.450  0.007  0.144  0.010 
+     0.9095  0.444  0.007  0.141  0.010 
+     0.9295  0.424  0.007  0.145  0.011 
+     0.9495  0.394  0.008  0.153  0.011 
+     0.9695  0.410  0.007  0.149  0.010 
+     0.9895  0.409  0.007  0.155  0.010 
+     1.0095  0.389  0.007  0.158  0.011 
+     1.0295  0.397  0.007  0.151  0.010 
+     1.0495  0.383  0.007  0.154  0.010 
+     1.0695  0.373  0.007  0.166  0.010 
+     1.0895  0.362  0.007  0.172  0.011 
+     1.1095  0.368  0.007  0.168  0.010 
+     1.1295  0.374  0.007  0.168  0.010 
+     1.1495  0.359  0.007  0.177  0.010 
+     1.1695  0.358  0.007  0.179  0.010 
+     1.1895  0.349  0.007  0.180  0.011 
+     1.2095  0.344  0.007  0.175  0.011 
+     1.2295  0.339  0.008  0.193  0.011 
+     1.2495  0.337  0.007  0.199  0.011 
+     1.2695  0.331  0.008  0.203  0.011 
+     1.2895  0.325  0.008  0.210  0.011 
+     1.3095  0.318  0.008  0.217  0.012 
+     1.3295  0.306  0.008  0.214  0.012 
+     1.3495  0.301  0.009  0.238  0.013 
+     1.3695  0.295  0.009  0.247  0.013 
+     1.3895  0.293  0.009  0.244  0.013 
+     1.4095  0.292  0.009  0.223  0.012 
+     1.4295  0.269  0.010  0.266  0.015 
+     1.4495  0.243  0.012  0.288  0.017 
+     1.4695  0.245  0.012  0.289  0.017 
+     1.4895  0.216  0.013  0.300  0.019 
+     1.5095  0.214  0.014  0.317  0.020 
+     1.5295  0.200  0.015  0.319  0.021 
+     1.5495  0.175  0.016  0.341  0.024 
+     1.5695  0.175  0.016  0.332  0.023 
+     1.5895  0.148  0.018  0.342  0.026 
+     1.6095  0.134  0.018  0.345  0.026 
+     1.6295  0.141  0.017  0.328  0.025 
+     1.6495  0.136  0.017  0.324  0.025 
+     1.6695  0.134  0.018  0.336  0.026 
+     1.6895  0.129  0.018  0.333  0.025 
+     1.7095  0.130  0.018  0.324  0.025 
+     1.7295  0.136  0.017  0.320  0.025 
+     1.7495  0.130  0.017  0.322  0.025 
+     1.7695  0.134  0.017  0.308  0.024 
+     1.7895  0.137  0.016  0.308  0.024 
+     1.8095  0.139  0.016  0.312  0.024 
+     1.8295  0.153  0.015  0.314  0.022 
+     1.8495  0.138  0.016  0.313  0.023 
+     1.8695  0.116  0.017  0.305  0.024 
+     1.8895  0.131  0.016  0.301  0.023 
+     1.9095  0.130  0.016  0.296  0.023 
+     1.9295  0.129  0.016  0.303  0.023 
+     1.9495  0.133  0.015  0.288  0.022 
+     1.9695  0.142  0.015  0.272  0.021 
+     1.9895  0.133  0.016  0.293  0.022 
+     2.0095  0.129  0.015  0.290  0.022 
+     2.0295  0.133  0.015  0.288  0.022 
+     2.0495  0.129  0.015  0.284  0.022 
+     2.0695  0.133  0.015  0.273  0.021 
+     2.0895  0.129  0.015  0.270  0.021 
+     2.1095  0.142  0.014  0.265  0.019 
+     2.1295  0.136  0.014  0.269  0.020 
+     2.1495  0.131  0.014  0.273  0.021 
+     2.1695  0.134  0.014  0.266  0.020 
+     2.1895  0.134  0.014  0.268  0.020 
+     2.2095  0.137  0.013  0.254  0.019 
+     2.2295  0.134  0.014  0.261  0.020 
+     2.2495  0.126  0.014  0.254  0.019 
+     2.2695  0.126  0.014  0.263  0.020 
+     2.2895  0.133  0.013  0.246  0.018 
+     2.3095  0.128  0.014  0.259  0.020 
+     2.3295  0.133  0.012  0.218  0.017 
+     2.3495  0.132  0.013  0.232  0.018 
+     2.3695  0.131  0.013  0.246  0.019 
+     2.3895  0.129  0.013  0.240  0.018 
+     2.4095  0.147  0.012  0.244  0.017 
+     2.4295  0.129  0.013  0.245  0.019 
+     2.4495  0.129  0.013  0.245  0.019 
+     2.4695  0.151  0.012  0.231  0.017 
+     2.4895  0.146  0.013  0.252  0.019 
+
